refactor: extracted cache code as a module
